% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airbnb.R
\name{airbnb_listing}
\alias{airbnb_listing}
\title{List the aibnb housing options from data based on the number of bedrooms
num_bedrooms that fall with budget with given range.}
\usage{
airbnb_listing(data, num_bedrooms, budget, range = budget * 0.2)
}
\arguments{
\item{data}{the airbnb dataset from a given city}

\item{num_bedrooms}{Number of bedrooms}

\item{budget}{Budget in currency from data}

\item{range}{Range for prices}
}
\value{
A data frame containing the bedrooms under the description's criteria
}
\description{
List the aibnb housing options from data based on the number of bedrooms
num_bedrooms that fall with budget with given range.
}
\examples{
airbnb_listing(amsterdam_airbnb, 2, 100, 10)
}
