% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airbnb_data.R
\docType{data}
\name{amsterdam}
\alias{amsterdam}
\title{This is data to be included in my package}
\format{A data frame with 53940 rows and 10 variables}
\usage{
data(amsterdam)
}
\description{
This is data to be included in my package
}
\keyword{data}
