% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airbnb.R
\name{neighborhood_average}
\alias{neighborhood_average}
\title{Calculate the average of the airbnb housing options per neighbourhood}
\usage{
neighborhood_average(city_data)
}
\arguments{
\item{city_data}{the airbnb dataset from a given city}
}
\value{
A data frame containing the price average by neighbourhood
}
\description{
Calculate the average of the airbnb housing options per neighbourhood
}
